+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/jvm/features/compiler/GenerateMetaRuntimeClasspathProviderTask.kt
@@ -1,13 +1,11 @@
 package com.javiersc.hubdle.project.extensions.kotlin.jvm.features.compiler
 
-import com.javiersc.gradle.properties.extensions.listProperty
-import com.javiersc.gradle.properties.extensions.property
+import com.javiersc.hubdle.project.extensions._internal.allKotlinSrcDirsWithoutBuild
+import com.javiersc.hubdle.project.extensions._internal.kotlinGeneratedTestSrcDirs
 import com.javiersc.hubdle.project.extensions.kotlin.shared.moduleAsString
 import com.javiersc.kotlin.stdlib.isNotNullNorBlank
-import io.gitlab.arturbosch.detekt.Detekt
 import java.io.File
 import javax.inject.Inject
-import org.gradle.api.DefaultTask
 import org.gradle.api.Project
 import org.gradle.api.artifacts.MinimalExternalModuleDependency
 import org.gradle.api.artifacts.ProjectDependency
@@ -15,46 +13,34 @@ import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.ListProperty
 import org.gradle.api.provider.Property
+import org.gradle.api.provider.Provider
 import org.gradle.api.provider.SetProperty
 import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.OutputFile
+import org.gradle.api.tasks.SourceTask
 import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.kotlin.dsl.get
 import org.gradle.kotlin.dsl.listProperty
-import org.gradle.kotlin.dsl.named
 import org.gradle.kotlin.dsl.property
 import org.gradle.kotlin.dsl.register
-import org.gradle.kotlin.dsl.the
-import org.gradle.kotlin.dsl.withType
-import org.jetbrains.kotlin.gradle.dsl.KotlinProjectExtension
-import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
 
 @CacheableTask
 public abstract class GenerateMetaRuntimeClasspathProviderTask
 @Inject
 constructor(
-    generatedTestSourceSetDir: Property<String>,
-    mainClass: Property<String>,
-    testDependenciesJarPaths: ListProperty<String>,
-    testProjectsJarPaths: ListProperty<String>,
     objects: ObjectFactory,
-) : DefaultTask() {
+) : SourceTask() {
+
+    @Input public val generatedTestSourceSetDir: Property<String> = objects.property<String>()
+
+    @Input public val mainClass: Property<String> = objects.property<String>()
 
     @Input
-    public val generatedTestSourceSetDir: Property<String> =
-        objects.property<String>().convention(generatedTestSourceSetDir)
+    public val testDependenciesJarPaths: ListProperty<String> = objects.listProperty<String>()
 
-    @Input public val mainClass: Property<String> = objects.property<String>().convention(mainClass)
-
-    @Input
-    public val testDependenciesJarPaths: ListProperty<String> =
-        objects.listProperty<String>().convention(testDependenciesJarPaths)
-
-    @Input
-    public val testProjectsJarPaths: ListProperty<String> =
-        objects.listProperty<String>().convention(testProjectsJarPaths)
+    @Input public val testProjectsJarPaths: ListProperty<String> = objects.listProperty<String>()
 
     @OutputFile
     public val generatedContentFile: RegularFileProperty =
@@ -115,20 +101,11 @@ constructor(
             testDependencies: SetProperty<MinimalExternalModuleDependency>,
             testProjects: SetProperty<ProjectDependency>,
         ): TaskProvider<GenerateMetaRuntimeClasspathProviderTask> {
-            val generatedTestSourceSetDir: Property<String> =
-                project.property {
-                    val testSet: KotlinSourceSet =
-                        project
-                            .the<KotlinProjectExtension>()
-                            .sourceSets
-                            .named<KotlinSourceSet>("test")
-                            .get()
-                    testSet.kotlin.srcDirs.firstOrNull { it.path.contains("generated") }?.path
-                        ?: project.buildDir.resolve("generated/test/kotlin").path
-                }
+            val generatedTestSourceSetDir: Provider<String> =
+                project.provider { project.kotlinGeneratedTestSrcDirs.get().first().path }
 
-            val testDependenciesJarPaths: ListProperty<String> =
-                project.listProperty {
+            val testDependenciesJarPaths: Provider<List<String>> =
+                project.provider {
                     val artifacts =
                         project.configurations["runtimeClasspath"]
                             .resolvedConfiguration
@@ -147,8 +124,8 @@ constructor(
                         .map { it.file.path }
                 }
 
-            val testProjectsJarPaths: ListProperty<String> =
-                project.listProperty {
+            val testProjectsJarPaths: Provider<List<String>> =
+                project.provider {
                     testProjects
                         .get()
                         .flatMap { projectDependency ->
@@ -168,19 +145,16 @@ constructor(
                         .map { it.path }
                 }
 
-            return project.tasks
-                .register<GenerateMetaRuntimeClasspathProviderTask>(
-                    NAME,
-                    generatedTestSourceSetDir,
-                    mainClass,
-                    testDependenciesJarPaths,
-                    testProjectsJarPaths,
-                )
-                .apply {
-                    project.tasks.withType<Detekt>().configureEach { detekt ->
-                        detekt.mustRunAfter(this)
-                    }
-                }
+            val generateMetaRuntimeClasspathProviderTask =
+                project.tasks.register<GenerateMetaRuntimeClasspathProviderTask>(NAME)
+            generateMetaRuntimeClasspathProviderTask.configure {
+                it.generatedTestSourceSetDir.convention(generatedTestSourceSetDir)
+                it.mainClass.convention(mainClass)
+                it.testDependenciesJarPaths.convention(testDependenciesJarPaths)
+                it.testProjectsJarPaths.convention(testProjectsJarPaths)
+                it.source(project.allKotlinSrcDirsWithoutBuild)
+            }
+            return generateMetaRuntimeClasspathProviderTask
         }
     }
 }