+++ CHANGELOG.md
@@ -12,10 +12,14 @@
 
 ### Fixed
 
+### Updated
+
+## [0.5.0-alpha.30] - 2023-06-08
+
+### Fixed
+
 - `GenerateProjectDataTask` dependency
 - `GenerateMetaRuntimeClasspathProviderTask` dependency
-
-### Updated
 
 ## [0.5.0-alpha.29] - 2023-06-08
 
@@ -393,7 +397,9 @@
 - `org.jetbrains.intellij.plugins:gradle-intellij-plugin -> 1.13.2`
 - `app.cash.molecule:molecule-gradle-plugin -> 0.8.0`
 
-[Unreleased]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-alpha.29...HEAD
+[Unreleased]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-alpha.30...HEAD
+
+[0.5.0-alpha.30]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-alpha.29...0.5.0-alpha.30
 
 [0.5.0-alpha.29]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-alpha.28...0.5.0-alpha.29